move manifest verification to separate module
