--- runtime_builders/config/runtimes.yaml
@@ -1,30 +1,0 @@
-schema_version: 1
-
-runtimes:
-  aspnetcore:
-    target:
-      file: aspnetcore-default-builder-20170822052732.yaml
-  go:
-    target:
-      runtime: go1.9
-  go1.9:
-    target:
-      file: go-1.9-builder-20170830135152.yaml
-  go1.8:
-    target:
-      file: go-1.8-builder-20170830135146.yaml
-  java:
-    target:
-      file: java-default-builder-20170829145312.yaml
-  nodejs:
-    target:
-      file: nodejs-default-builder-20170824135451.yaml
-  php:
-    target:
-      file: php-default-builder-20170823151341.yaml
-  python:
-    target:
-      file: python-default-builder-20170825144546.yaml.yaml
-  ruby:
-    target:
-      file: ruby-default-builder-20170830110745.yaml
