+++ runtime_builders/publish_builders.py
@@ -22,6 +22,7 @@ import yaml
 import sys
 
 import builder_util
+import verify_manifest
 
 
 def main():
@@ -41,7 +42,7 @@ def main():
             logging.error('Please provide path to runtime.yaml manifest.')
         with open(args.manifest, 'r') as f:
             manifest = yaml.safe_load(f)
-        builder_util.verify_manifest(manifest)
+        verify_manifest.verify_manifest(manifest)
         builder_util.copy_to_gcs(args.manifest, builder_util.MANIFEST_FILE)
 
         if args.directory:
